* libtool: Update to libtool 1.2.4.
* magick/magick.h (MagickInfo): Store string attributes as 'const
char *' to statically allocated data rather than as heap allocated
strings.  THIS IS AN INTERFACE CHANGE.  Using a coder written to
the old interface is non-fatal but may resemble a small memory
leak.  The reason for the change is to avoid at least 880 needless
malloc()/strlen()/strlcpy() operations at initialization time, and
at least 880 free() calls at destruction time.  While these
operations did not take long, they are still an unnecessary
overhead, which is increased in thread-safe applications.
* magick/module.c (OpenModule): Ignore requests to open modules
* GraphicsMagick.spec.in: Add --with-included-ltdl to the default
options since this seems safest until the libltdl validation logic
is fully robust.  This should be made user-configurable in the
future.
* magick/Makefile.am (magick_libGraphicsMagick_la_LIBADD): Apply
libltdl dependency argument as required for building.
* configure.ac: Intuit if the GraphicsMagick library will depend
on -lltdl.
* Magick++/bin/GraphicsMagick++-config.in: Use substitutions
rather than invoking GraphicsMagick-config in order to determine
GraphicsMagick library usage requirements.  This avoids problems
when GraphicsMagick-config is not in the executable search path.


* GraphicsMagick.spec.in: NEWS was renamed to NEWS.txt.  Enable
libtool verbose output so it is possible to diagnose build
failures.